moved validation model to correct location. fixed linting issues
