+++ src/app/Model/ValidationError.cs
@@ -1,6 +1,6 @@
 ﻿using System.Text.Json.Serialization;
 
-namespace CSE.Helium.Validation
+namespace CSE.Helium.Model
 {
     public class ValidationError
     {
