+++ src/app/Validation/ValidationProblemDetailsResult.cs
@@ -1,4 +1,5 @@
 ﻿using Helium;
+using CSE.Helium.Model;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.DependencyInjection;
